Improve code standard
